refactor(S2/批8): interp 裸指针入槽显式化（`as int`）——S1P 总点 206→200、类3 `w64` 9→3
`src/compiler/interp.cr`：`bp`/`bp2 := alloc(...)` 按模型是 **string**，但其值语义是**裸指针**
（后续经 `r64` 读出做 `load8/store8` 与指针算术）⇒ 入值槽处补 **显式** `as int`（6 处：
`IR_ALLOC`/`IR_MAKE_STRUCT`/`IR_MAKE_ENUM`/`IR_MAKE_ARRAY` 等的分配结果写入）。

**收敛读数**（冷缓存、逐轮）：S1P 总打点 206 → **200**；`int←string` 类 9 → **3**；
S2 两类合计 **68 → 29**（余量全在 interp 值槽的读出口：`r64` 读出后当串用 —— 需按**被调签名语义**
逐点分类：`istr_get` 下标→串 / `as string` 裸指针 / `str_intern` 串→下标）。

⚠ 同批已记：一次误改（把返回 `int` 的 `ir_interp_str_*` 包 `str_intern`）被 oracle 当场抓出并回退 ✓。

+++ src/compiler/interp.cr
@@ -286,7 +286,7 @@ fn ir_interp_run_fn(cfi: int, arg_base: int, argc: int) -> int {
             bp2 := alloc(need3);
             vi3 : ., mut = 0;
             loop { if vi3 >= need3 { break; } store8(bp2, vi3, 0); vi3 = vi3 + 1; }
-            w64(g_ir_vals, d2 * 8, bp2);
+            w64(g_ir_vals, d2 * 8, bp2 as int);   // 值槽存**裸指针**（alloc 结果按模型是 string）⇒ 显式
         }
         if op2 == 13 && d2 >= 0 && t1 >= 0 {  // IR_LOAD_INDEX: t1=arr_var, t3=literal_idx
             av2 := r64(g_ir_vals, t1 * 8);
@@ -331,7 +331,7 @@ fn ir_interp_run_fn(cfi: int, arg_base: int, argc: int) -> int {
             bp2 := alloc(need3);
             vi3 : ., mut = 0;
             loop { if vi3 >= need3 { break; } store8(bp2, vi3, 0); vi3 = vi3 + 1; }
-            w64(g_ir_vals, d2 * 8, bp2);
+            w64(g_ir_vals, d2 * 8, bp2 as int);   // 值槽存**裸指针**（alloc 结果按模型是 string）⇒ 显式
         }
         if op2 == 32 && d2 >= 0 { w64(g_ir_vals, d2 * 8, 0); }  // IR_ARENA_NEW
         if op2 == 33 { }  // IR_ARENA_RESET
@@ -348,7 +348,7 @@ fn ir_interp_run_fn(cfi: int, arg_base: int, argc: int) -> int {
                 vi3 : ., mut = 0;
                 loop { if vi3 >= need3 { break; } store8(bp2, vi3, 0); vi3 = vi3 + 1; }
                 w64(bp2, 0, t1);
-                w64(g_ir_vals, d2 * 8, bp2);
+                w64(g_ir_vals, d2 * 8, bp2 as int);   // 值槽存**裸指针**（alloc 结果按模型是 string）⇒ 显式
             }
         }
         // IR_LOAD_ENUM_TAG (23)：d := M[ρ(s1)+0]
@@ -573,7 +573,7 @@ fn ir_interpret() -> int {
                 bp := alloc(need2);
                 vi2 : ., mut = 0;
                 loop { if vi2 >= need2 { break; } store8(bp, vi2, 0); vi2 = vi2 + 1; }
-                w64(g_ir_vals, d * 8, bp);
+                w64(g_ir_vals, d * 8, bp as int);   // 同上：裸指针入槽（显式）
             }
         }
         if op == 8 {  // IR_ALLOC_ARRAY
@@ -584,7 +584,7 @@ fn ir_interpret() -> int {
                 bp := alloc(need2);
                 vi2 : ., mut = 0;
                 loop { if vi2 >= need2 { break; } store8(bp, vi2, 0); vi2 = vi2 + 1; }
-                w64(g_ir_vals, d * 8, bp);
+                w64(g_ir_vals, d * 8, bp as int);   // 同上：裸指针入槽（显式）
             }
         }
         if op == 9 { if s1 >= 0 && s2 >= 0 { w64(g_ir_vals, s1 * 8, r64(g_ir_vals, s2 * 8)); } }  // IR_STORE
@@ -642,7 +642,7 @@ fn ir_interpret() -> int {
                 vi2 : ., mut = 0;
                 loop { if vi2 >= need2 { break; } store8(bp, vi2, 0); vi2 = vi2 + 1; }
                 w64(bp, 0, s1);
-                w64(g_ir_vals, d * 8, bp);
+                w64(g_ir_vals, d * 8, bp as int);   // 同上：裸指针入槽（显式）
             }
         }
         // IR_LOAD_ENUM_TAG (23)：d := M[ρ(s1)+0]
